move email logic into editprofile controller and remove email controller
