+++ identity/app/views/fragments/profile/privacyForm.scala.html
@@ -55,21 +55,6 @@
         </div>
     }
 
-    @* Newsletters *@
-    <fieldset class="fieldset">
-        <div class="fieldset__heading">
-            <h2 class="form__heading">Newsletters</h2>
-        </div>
-
-        <div class="fieldset__fields">
-            <ul class="u-unstyled">
-                <li class="form-field">
-                    <label class="label">Click <a href=@idUrlBuilder.buildUrl("/email-prefs")>here</a> to manage your email subscriptions to editorial newsletters.</label>
-                </li>
-            </ul>
-        </div>
-    </fieldset>
-
     @* Marketing Consent *@
     <fieldset class="fieldset">
