+++ identity/app/views/profile/emailPrefs.scala.html
@@ -60,7 +60,7 @@
 }
 
     <div class="identity-wrapper monocolumn-wrapper">
-        <h1 class="identity-title">Email preferences</h1>
+        <h1 class="identity-title">Editorial Newsletters</h1>
         <h4 class="email-subscription__meta">To change your email address please click <a href=@buildIdentityUrl("account/edit")>here.</a></h4>
         <form class="form" novalidate action="@buildIdentityUrl("email-prefs")" role="main" method="post">
             @views.html.helper.CSRF.formField
