+++ identity/conf/routes
@@ -17,8 +17,7 @@ POST        /password/change                        controllers.ChangePasswordCo
 GET         /password/reset-confirmation            controllers.ResetPasswordController.renderPasswordResetConfirmation
 GET         /password/email-sent                    controllers.ResetPasswordController.renderEmailSentConfirmation
 GET         /requestnewtoken                        controllers.ResetPasswordController.requestNewToken
-GET         /email-prefs                            controllers.EmailController.preferences
-POST        /email-prefs                            controllers.EmailController.savePreferences
+POST        /email-prefs                            controllers.EditProfileController.saveEmailPreferences
 GET         /user/id/:id                            controllers.PublicProfileController.renderProfileFromId(id: String, activityType = "discussions")
 GET         /user/id/:id/:activityType              controllers.PublicProfileController.renderProfileFromId(id: String, activityType: String)
 GET         /user/:vanityUrl                        controllers.PublicProfileController.renderProfileFromVanityUrl(vanityUrl: String, activityType = "discussions")
